Upgraded package versions, made changes to accomodate the upgrade

+++ pi_server/src/cloudflared.rs
@@ -74,7 +74,7 @@ pub async fn init_cloudflare<'a>() -> anyhow::Result<Cloudflared<'a>> {
 
 pub struct Cloudflared<'a> {
     cmd: &'a str,
-    port: u32,
+    port: u16,
 }
 
 impl<'a> Cloudflared<'a> {
